pyproject.toml
[tool.poetry]
name = "synthetic_population_schema"
version = "0.0.0.post15.dev0+769a593"
description = "A LinkML schema capturing the model used to enumerate https://github.com/RTIInternational/SyntheticPopulations?tab=readme-ov-file#data-identification-and-metadata"
authors = ["Sierra Moxon <smoxon@lbl.gov>"]
license = "MIT"
readme = "README.md"
include = ["README.md", "src/synthetic_population_schema/schema", "project"]

[tool.poetry.dependencies]
python = "^3.9"
linkml-runtime = "^1.8.3"
linkml-transformer = "^0.1"
linkml = "^1.8.1"
gitpython = "^3.1.44"
pystow = "^0.7.0"
pyneo = "^0.2.0"

[tool.poetry.group.dev.dependencies]
linkml-renderer = "^0.1.2"

[tool.poetry-dynamic-versioning]
enable = false
vcs = "git"
style = "pep440"

[tool.poetry.dev-dependencies]
linkml = "^1.3.5"
mkdocs-material = "*"
mkdocs-mermaid2-plugin = "^1.1.1"
kghub-downloader = "*"

[build-system]
requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning"]
build-backend = "poetry_dynamic_versioning.backend"

[tool.poetry.extras]
docs = ["linkml", "mkdocs-material"]

[scripts]
load_neo = "src.synthetic_population_schema.scripts.load_neo:main"